client-python/pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "a2a-registry-client"
version = "0.1.0"
description = "Python client library for the A2A Registry"
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}
authors = [
    {name = "A2A Registry Community", email = "contact@a2aregistry.org"}
]
keywords = ["a2a", "agents", "ai", "registry", "protocol"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]

dependencies = [
    "requests>=2.31.0",
    "pydantic>=2.0.0",
    "fastmcp>=2.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "pytest-asyncio>=0.21.0",
    "black>=23.0.0",
    "ty>=0.0.1a1",
    "ruff>=0.1.0",
]
async = [
    "aiohttp>=3.8.0",
]
a2a = [
    "a2a-sdk>=0.3.7",
]
all = [
    "aiohttp>=3.8.0",
    "a2a-sdk>=0.3.7",
]

[project.urls]
Homepage = "https://www.a2aregistry.org"
Documentation = "https://github.com/prassanna-ravishankar/a2a-registry"
Repository = "https://github.com/prassanna-ravishankar/a2a-registry"
Issues = "https://github.com/prassanna-ravishankar/a2a-registry/issues"

[project.scripts]
a2a-registry-client = "a2a_registry.mcp_server:main"

[tool.setuptools.packages.find]
where = ["src"]

[tool.black]
line-length = 100
target-version = ["py310", "py311", "py312"]

[tool.ruff]
line-length = 100
target-version = "py310"
select = ["E", "F", "I", "N", "W"]

